Moved routes out into separate helper, and applied alternative camelCasing rules.
