+++ assets/pages/monthly-contributions-existing/monthlyContributionsExisting.jsx
@@ -11,7 +11,7 @@ import CtaLink from 'components/ctaLink/ctaLink';
 import InfoSection from 'components/infoSection/infoSection';
 
 import pageStartup from 'helpers/pageStartup';
-import { routes } from 'helpers/url';
+import { routes } from 'helpers/routes';
 
 
 // ----- Page Startup ----- //
@@ -35,7 +35,7 @@ const content = (
           </h2>
           <CtaLink
             text="Make a one-off contribution"
-            url={routes.oneoffContribCheckout}
+            url={routes.oneOffContribCheckout}
           />
         </div>
         <InfoSection heading="Questions?" className="existing__questions">
